requirements.txt
google-genai
jinja2
python-dotenv
yt-dlp
pytest
